Updated delete user case

+++ sas/booking/views.py
@@ -19,4 +19,10 @@ def new_user(request):
     return render(request, 'booking/newUser.html', {'form_user':form})
 
 def delete_user(request, key):
-    User.objects.get(key = key).delete()
+     if request.POST['delete']:
+         User.objects.get(key = key).delete()
+         return render(request, 'booking/index.html', {})
+     elif request.POST['cancel']:
+        return render(request, 'booking/index.html',{})
+     else:
+        return render(request, 'booking/index.html',{})
